refactor(plugins): remove redundant catalog builders and private barrels
