--- extensions/qa-lab/src/protocol.ts
@@ -1,22 +1,0 @@
-// Qa Lab plugin module implements protocol behavior.
-export type {
-  QaBusAttachment,
-  QaBusConversation,
-  QaBusCreateThreadInput,
-  QaBusDeleteMessageInput,
-  QaBusEditMessageInput,
-  QaBusEvent,
-  QaBusInboundMessageInput,
-  QaBusMessage,
-  QaBusOutboundMessageInput,
-  QaBusPollInput,
-  QaBusPollResult,
-  QaBusReactToMessageInput,
-  QaBusReadMessageInput,
-  QaBusSearchMessagesInput,
-  QaBusSnapshotConversation,
-  QaBusStateSnapshot,
-  QaBusThread,
-  QaBusToolCall,
-  QaBusWaitForInput,
-} from "openclaw/plugin-sdk/qa-channel-protocol";
