+++ extensions/qa-lab/src/runtime-api.ts
@@ -42,4 +42,4 @@ export type {
   QaBusThread,
   QaBusToolCall,
   QaBusWaitForInput,
-} from "./protocol.js";
+} from "openclaw/plugin-sdk/qa-channel-protocol";
