--- extensions/telegram/src/lane-delivery.ts
@@ -1,12 +1,0 @@
-// Telegram plugin module implements lane delivery behavior.
-export {
-  createLaneTextDeliverer,
-  type LaneTextDeliverer,
-  type DraftLaneState,
-  type LaneDeliveryResult,
-  type LaneName,
-} from "./lane-delivery-text-deliverer.js";
-export {
-  createLaneDeliveryStateTracker,
-  type LaneDeliveryStateTracker,
-} from "./lane-delivery-state.js";
